Created new package multi_astra_camera
