+++ astra_nodelets.xml
@@ -1,6 +1,6 @@
-<library path="lib/libastra_camera_nodelet">
+<library path="lib/libmulti_astra_camera_nodelet">
 
-  <class name="astra_camera/AstraDriverNodelet" type="astra_camera::AstraDriverNodelet" base_class_type="nodelet::Nodelet">
+  <class name="multi_astra_camera/AstraDriverNodelet" type="multi_astra_camera::AstraDriverNodelet" base_class_type="nodelet::Nodelet">
     <description>
       Astra camera driver nodelet.
     </description>
