+++ cfg/Astra.cfg
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-PACKAGE='astra_camera'
+PACKAGE='multi_astra_camera'
 
 from dynamic_reconfigure.parameter_generator_catkin import *
