+++ include/multi_astra_camera/COBDevice.h
