+++ include/multi_astra_camera/astra_device.h
@@ -35,9 +35,9 @@
 
 #include "openni2/OpenNI.h"
 
-#include "astra_camera/astra_video_mode.h"
+#include "multi_astra_camera/astra_video_mode.h"
 
-#include "astra_camera/astra_exception.h"
+#include "multi_astra_camera/astra_exception.h"
 
 #include <boost/shared_ptr.hpp>
 #include <boost/cstdint.hpp>
