+++ include/multi_astra_camera/astra_device_info.h
