+++ include/multi_astra_camera/astra_frame_reader.h
@@ -1,9 +1,9 @@
 #ifndef ASTRA_FRAME_READER_H_
 #define ASTRA_FRAME_READER_H_ 
 
-#include "astra_camera/astra_device.h"
+#include "multi_astra_camera/astra_device.h"
 #include "openni2/OpenNI.h"
-#include "astra_camera/COBDevice.h"
+#include "multi_astra_camera/COBDevice.h"
 
 #include <ros/ros.h>
 #include <sensor_msgs/Image.h>
