+++ include/multi_astra_camera/astra_multi_driver.h
@@ -1,6 +1,11 @@
 #ifndef ASTRA_MULTI_DRIVER_H
 #define ASTRA_MULTI_DRIVER_H
 
+#include <ros/ros.h>
+
+#include <string>
+#include <vector>
+#include <map>
 #include <boost/shared_ptr.hpp>
 #include <boost/cstdint.hpp>
 #include <boost/bind.hpp>
@@ -8,28 +13,20 @@
 
 #include <sensor_msgs/Image.h>
 #include <std_srvs/SetBool.h>
-
 #include <dynamic_reconfigure/server.h>
-#include <astra_camera/AstraConfig.h>
-
 #include <image_transport/image_transport.h>
 #include <camera_info_manager/camera_info_manager.h>
 
-#include <string>
-#include <vector>
+#include <multi_astra_camera/AstraConfig.h>
+#include "multi_astra_camera/astra_device_manager.h"
+#include "multi_astra_camera/astra_device.h"
+#include "multi_astra_camera/astra_advanced_device.h"
+#include "multi_astra_camera/astra_video_mode.h"
+#include "multi_astra_camera/GetSerial.h"
 
-#include "astra_camera/astra_device_manager.h"
-#include "astra_camera/astra_device.h"
-#include "astra_camera/astra_advanced_device.h"
-#include "astra_camera/astra_video_mode.h"
-#include "astra_camera/GetSerial.h"
-
-#include <ros/ros.h>
-
-#include "astra_camera/astra_device_manager.h"
-#include "astra_camera/astra_driver.h"
-#include "astra_camera/astra_registration_info.h"
-#include <map>
+#include "multi_astra_camera/astra_device_manager.h"
+#include "multi_astra_camera/astra_driver.h"
+#include "multi_astra_camera/astra_registration_info.h"
 
 namespace astra_wrapper
 {
@@ -48,7 +45,7 @@ private:
 
   ros::Subscriber astra_register_sub_;
 
-  void RegistrationCallback(const astra_camera::astra_registration_info::ConstPtr& msg);
+  void RegistrationCallback(const multi_astra_camera::astra_registration_info::ConstPtr& msg);
 };
 
 }
