+++ include/multi_astra_camera/astra_video_mode.h
