+++ scripts/create_udev_rules
@@ -5,7 +5,7 @@ echo "This script copies a udev rule to /etc to facilitate bringing"
 echo "up the astra usb connection as /dev/astra*"
 echo ""
 
-sudo cp `rospack find astra_camera`/56-orbbec-usb.rules /etc/udev/rules.d
+sudo cp `rospack find multi_astra_camera`/56-orbbec-usb.rules /etc/udev/rules.d
 
 
 echo ""
