+++ src/astra_device_manager.cpp
@@ -30,11 +30,11 @@
  *      Author: Tim Liu (liuhua@orbbec.com)
  */
 
-#include "astra_camera/astra_device_manager.h"
-#include "astra_camera/astra_convert.h"
-#include "astra_camera/astra_device.h"
-#include "astra_camera/astra_advanced_device.h"
-#include "astra_camera/astra_exception.h"
+#include "multi_astra_camera/astra_device_manager.h"
+#include "multi_astra_camera/astra_convert.h"
+#include "multi_astra_camera/astra_device.h"
+#include "multi_astra_camera/astra_advanced_device.h"
+#include "multi_astra_camera/astra_exception.h"
 
 #include <boost/make_shared.hpp>
