+++ src/astra_multi_driver.cpp
@@ -1,4 +1,4 @@
-#include "astra_camera/astra_multi_driver.h"
+#include "multi_astra_camera/astra_multi_driver.h"
 #include <boost/make_shared.hpp>
 
 namespace astra_wrapper
@@ -14,7 +14,7 @@ AstraMultiDriver::AstraMultiDriver(ros::NodeHandle& n, ros::NodeHandle& pnh) :
   astra_register_sub_ = nh_.subscribe("/astra_registration", 1, &AstraMultiDriver::RegistrationCallback, this);
 }
 
-void AstraMultiDriver::RegistrationCallback(const astra_camera::astra_registration_info::ConstPtr& msg)
+void AstraMultiDriver::RegistrationCallback(const multi_astra_camera::astra_registration_info::ConstPtr& msg)
 {
   ROS_INFO_STREAM(GetLogPrefix("AstraMultiDriver", msg->ns) << "serial_no: " << msg->serial_no << ", is_advanced: " << (int)msg->is_advanced);
