+++ src/astra_video_mode.cpp
@@ -30,7 +30,7 @@
  *      Author: Tim Liu (liuhua@orbbec.com)
  */
 
-#include "astra_camera/astra_video_mode.h"
+#include "multi_astra_camera/astra_video_mode.h"
 
 namespace astra_wrapper
 {
